# test-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation with Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get list of schedules",
                "operationId": "c2791edf6c0550c67fac652cad98e09e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Schedule"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Create a new schedule",
                "operationId": "cec681726db3f091faae4410db8ee14c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Schedule"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/get_all_by_user/list": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Get list of schedules by logged user",
                "operationId": "e8f978d4e3a1d5dc2704ffc116256ce9",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Schedule"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "properties": {
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "token": {
                        "type": "string",
                        "example": "a-long-jwt-token"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Schedule": {
                "title": "Schedule",
                "required": [
                    "title",
                    "type",
                    "description",
                    "status",
                    "start_at",
                    "conclusion_at",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Meeting"
                    },
                    "type": {
                        "type": "string",
                        "example": "meeting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Team meeting"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "start_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15T00:00:00Z"
                    },
                    "conclusion_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15T01:00:00Z"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "ScheduleStoreRequest": {
                "title": "Schedule Store Request",
                "required": [
                    "title",
                    "type",
                    "description",
                    "status",
                    "start_at",
                    "conclusion_at",
                    "user_id"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Meeting"
                    },
                    "type": {
                        "type": "string",
                        "example": "meeting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Team meeting"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "start_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15T00:00:00Z"
                    },
                    "conclusion_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-15T01:00:00Z"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Schedules",
            "description": "Schedules"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ]
}